+++ test/ruby_test.rb
@@ -165,9 +165,8 @@ class RubyTest < Minitest::Test
         assert_equal "1.32.0", hash["rubocop-rspec"]
       end
 
-      messages = trace_writer.writer.select { |m| m[:trace] == :message }.map { |m| m[:message] }
-      assert_equal(["Generating optimized Gemfile...", <<~MSG.strip, "Installing gems..."], messages)
-        ---
+      gemfile_content_log = <<~RUBY.strip
+        ### Auto-generated Gemfile ###
         source "https://rubygems.org"
 
         gem "strong_json", "0.5.0", "<= 0.8.0"
@@ -175,8 +174,11 @@ class RubyTest < Minitest::Test
         git "https://github.com/rubocop-hq/rubocop-rspec.git", tag: "v1.32.0" do
           gem "rubocop-rspec"
         end
-        ---
-      MSG
+      RUBY
+      assert system("ruby", "-ce", gemfile_content_log)
+
+      messages = trace_writer.writer.select { |m| m[:trace] == :message }.map { |m| m[:message] }
+      assert_equal(["Generating optimized Gemfile...", gemfile_content_log, "Installing gems..."], messages)
     end
   end